+++ apis/r/R/SOMANDArrayBase.R
@@ -41,7 +41,7 @@ SOMANDArrayBase <- R6::R6Class(
         stop(
           paste(
             strwrap(private$.internal_use_only("create", "collection")),
-            collapse = '\n'
+            collapse = "\n"
           ),
           call. = FALSE
         )
@@ -57,7 +57,6 @@ SOMANDArrayBase <- R6::R6Class(
       tiledb_create_options <- TileDBCreateOptions$new(platform_config)
 
       ## create array
-      # ctxptr <- self$tiledbsoma_ctx$context()
       sparse <- if (inherits(self, "SOMASparseNDArray")) {
         TRUE
       } else if (inherits(self, "SOMADenseNDArray")) {
@@ -80,28 +79,6 @@ SOMANDArrayBase <- R6::R6Class(
 
       self$reopen("WRITE", tiledb_timestamp = self$tiledb_timestamp)
       return(self)
-    },
-
-    ## needed eg after open() to set (Arrow) type
-    #' @description Sets a cache value for the datatype.
-    #'
-    #' @section Lifecycle:
-    #' As of \pkg{tiledbsoma} 2.1.0, \code{$set_data_type()} is deprecated; this
-    #' functionality is no longer needed as \code{libtiledbsoma} now accurately
-    #' sets the TileDB data type upon array opening
-    #'
-    #' @param type A character value describing the TileDB data type.
-    #'
-    set_data_type = function(type) {
-      .deprecate(
-        when = "2.1.0",
-        what = sprintf("%s$set_data_type()", class(self)[1L])
-      )
-      soma_debug(sprintf(
-        "[SOMANDArrayBase::set_data_type] caching type %s",
-        type$ToString()
-      ))
-      private$.type <- type
     },
 
     #' @description Test if the array has the upgraded resizeable shape feature