+++ apis/r/R/zzz.R
@@ -68,12 +68,9 @@ NULL
 #' @export
 #'
 soma_context <- function(config = NULL) {
-  .deprecate(
+  lifecycle::deprecate_stop(
     when = "2.3.0",
     what = "soma_context()",
     details = "Use new `SOMAContext` class instead."
   )
-  context <- SOMAContext$new(config)
-  .pkgenv[["somactx"]] <- context
-  return(SOMAContext$new(config))
 }